+++ hostfunctions/textgeneration.go
@@ -6,12 +6,13 @@ package hostfunctions
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 
 	"hmruntime/functions/assemblyscript"
 	"hmruntime/logger"
 	"hmruntime/models"
+	"hmruntime/models/openai"
+	"hmruntime/utils"
 
 	wasm "github.com/tetratelabs/wazero/api"
 )
@@ -51,7 +52,7 @@ func hostInvokeTextGenerator(ctx context.Context, mod wasm.Module, pModelName ui
 		// safeguard: test is the output is a valid json
 		// test every Choices.Message.Content
 		for _, choice := range result.Choices {
-			_, err := json.Marshal(choice.Message.Content)
+			_, err := utils.JsonSerialize(choice.Message.Content)
 			if err != nil {
 				logger.Err(ctx, err).Msg("One of the generated message is not a valid JSON.")
 				return 0