+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 2024-05-13 - Version 0.7.0
+
+- Sentry is no longer used when `HYPERMODE_DEBUG` is enabled [#187](https://github.com/gohypermode/runtime/pull/187)
+- Only listen on `localhost` when `HYPERMODE_DEBUG` is enabled, to prevent firewall prompt [#188](https://github.com/gohypermode/runtime/pull/188)
+- Improve support for marshaling classes [#189](https://github.com/gohypermode/runtime/pull/189) [#191](https://github.com/gohypermode/runtime/pull/191)
+- Add support for binary data fields [#190](https://github.com/gohypermode/runtime/pull/190)
+- Add host function for HTTP fetch [#191](https://github.com/gohypermode/runtime/pull/191)
+
+## 2024-05-08 - Version 0.6.6
+
+- Remove `Access-Control-Allow-Credentials`. Add `Access-Control-Request-Headers` [#180](https://github.com/gohypermode/runtime/pull/180)
+- Restrict incoming http requests methods [#182](https://github.com/gohypermode/runtime/pull/182)
+
+## 2024-05-07 - Version 0.6.5
+
+- Add `Access-Control-Allow-Credentials` in CORS preflight [#177](https://github.com/gohypermode/runtime/pull/177)
+
 ## 2024-05-03 - Version 0.6.4
 
 - Add CORS support to all endpoints [#171](https://github.com/gohypermode/runtime/pull/171)